pyproject.toml
[project]
name = "prompt-template"
version = "0.1.0"
description = ""
authors = [
    {name = "varun"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langchain-openai (>=0.3.33,<0.4.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "ollama (>=0.5.3,<0.6.0)",
    "langchain-ollama (>=0.3.8,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
